Updated jquery.shinybox.js
I have changed the calculation of width and height of shiny box as in mobiles the height of window will keep changing when the user scrolls(It would be reduced when toolbar is shown and increased when toolbar disappears. This happens on scroll of a page. So removed the constant height and used window.innerHeight. I have tested it in phones( Android device, Iphone and Ipad) and on pc(Chrome, safari(mac), firefox and IE) preview in both desktop and mobile view. It works fine. The problem with the previos code can be seen in issue https://jira.one.com/browse/ONEWEB-10139